Added Jasem's portrait to website
Jasem's portrait was added to the Images folder and is now displayed in the "E-Board" section of the website. To properly align the second row of the eboard with the first row, the main.css file was slightly modified. Additionally, Bana's portrait is now in the Images folder, but it is currently not being displayed on the website.